+++ addon-test-support/-private/converters.js
@@ -1,4 +1,4 @@
-import findByLabel from 'ember-cli-yadda-opinionated/test-support/-private/find-by-label';
+import { findByLabel } from 'ember-cli-yadda-opinionated/test-support/-private/dom-helpers';
 
 export function element(label, next) {
   const collectionOrElement = findByLabel(label);