+++ addon-test-support/index.js
@@ -6,7 +6,7 @@ import {
 } from 'ember-cli-yadda-opinionated/test-support/-private/regex';
 
 import { labelMap, transformsMap } from 'ember-cli-yadda-opinionated/test-support/-private/maps';
-import findBylabel from 'ember-cli-yadda-opinionated/test-support/-private/find-by-label';
+import { clickByLabel, fillInByLabel, findByLabel } from 'ember-cli-yadda-opinionated/test-support/-private/dom-helpers';
 import selectorFromLabel from 'ember-cli-yadda-opinionated/test-support/-private/selector-from-label';
 import setupDictionary from 'ember-cli-yadda-opinionated/test-support/-private/setup-dictionary';
 import givenSteps from 'ember-cli-yadda-opinionated/test-support/-private/given-steps';
@@ -16,9 +16,11 @@ import whenSteps from 'ember-cli-yadda-opinionated/test-support/-private/when-st
 const opinionatedSteps = {...givenSteps, ...whenSteps, ...thenSteps};
 
 export {
+  clickByLabel,
   composeSteps,
   element,
-  findBylabel,
+  fillInByLabel,
+  findByLabel,
   givenSteps,
   labelMap,
   opinionatedSteps,